More createTemporaryMessage to a store action
Since createTemporaryMessage heavily depends on a store, moved it into a
messagesStore action to decouple it from the global store module.

Note: even though it's synchronous, the dispatch() function returns a
promise so we need to treat it as if it was async.
