+++ src/store/fileUploadStore.js
@@ -26,7 +26,6 @@ import { showError } from '@nextcloud/dialogs'
 import fromStateOr from './helper'
 import { findUniquePath, getFileExtension } from '../utils/fileUpload'
 import moment from '@nextcloud/moment'
-import createTemporaryMessage from '../utils/temporaryMessage'
 import { EventBus } from '../services/EventBus'
 import { shareFile } from '../services/filesSharingServices'
 import { setAttachmentFolder } from '../services/settingsService'
@@ -195,19 +194,19 @@ const actions = {
 	 * @param {number} uploadId a unique id for the upload operation indexing
 	 * @param {bool} rename whether to rename the files (usually after pasting)
 	 */
-	initialiseUpload({ commit, dispatch }, { uploadId, token, files, rename = false }) {
-		if (rename) {
-			files.forEach(file => {
-				// note: can't overwrite the original read-only name attribute
-				file.newName = moment(file.lastModified || file.lastModifiedDate).format('YYYYMMDD_HHmmss')
-					+ getFileExtension(file.name)
-			})
-		}
-
+	async initialiseUpload({ commit, dispatch }, { uploadId, token, files, rename = false }) {
 		// Set last upload id
 		commit('setCurrentUploadId', uploadId)
 
-		files.forEach(file => {
+		for (let i = 0; i < files.length; i++) {
+			const file = files[i]
+
+			if (rename) {
+				// note: can't overwrite the original read-only name attribute
+				file.newName = moment(file.lastModified || file.lastModifiedDate).format('YYYYMMDD_HHmmss')
+					+ getFileExtension(file.name)
+			}
+
 			// Get localurl for some image previews
 			let localUrl = ''
 			if (file.type === 'image/png' || file.type === 'image/gif' || file.type === 'image/jpeg') {
@@ -219,10 +218,12 @@ const actions = {
 			const date = new Date()
 			const index = 'temp_' + date.getTime() + Math.random()
 			// Create temporary message for the file and add it to the message list
-			const temporaryMessage = createTemporaryMessage('{file}', token, uploadId, index, file, localUrl)
+			const temporaryMessage = await dispatch('createTemporaryMessage', {
+				text: '{file}', token, uploadId, index, file, localUrl,
+			})
 			console.debug('temporarymessage: ', temporaryMessage, 'uploadId', uploadId)
 			commit('addFileToBeUploaded', { file, temporaryMessage })
-		})
+		}
 	},
 
 	/**
